requirements.txt
nornir
nornir-netbox
requests
scapy
pyyaml
